requirements.txt
streamlit
pdf2image
pytesseract
Pillow
indic-transliteration
fpdf
python-magic
